code/leetcode/leetcode_java/src/main/java/org/sunrain/study/leetcode/Solution_0136.java
package org.sunrain.study.leetcode;

import java.util.HashMap;

/**
 * Created by qichao on 2019/11/16.
 * 给定一个非空整数数组，除了某个元素只出现一次以外，其余每个元素均出现两次。找出那个只出现了一次的元素。
 * <p>
 * 说明：
 * <p>
 * 你的算法应该具有线性时间复杂度。 你可以不使用额外空间来实现吗？
 * <p>
 * 示例 1:
 * <p>
 * 输入: [2,2,1]
 * 输出: 1
 * 示例 2:
 * <p>
 * 输入: [4,1,2,1,2]
 * 输出: 4
 * <p>
 * 来源：力扣（LeetCode）
 * 链接：https://leetcode-cn.com/problems/single-number
 * 著作权归领扣网络所有。商业转载请联系官方授权，非商业转载请注明出处。
 */
public class Solution_0136 {

    public int singleNumber(int[] nums) {

        if (nums == null) {
            return 0;
        }

        HashMap<Integer, Integer> map = new HashMap();
        for (int i = 0; i < nums.length; i++) {
            if (map.get(nums[i]) == null) {
                map.put(nums[i], 1);
                continue;
            }
            map.remove(nums[i]);
        }

        for (int key : map.keySet()) {
            return key;
        }

        return 0;
    }
}
